# Conflicts:
#	src/main/java/Main.java
#	src/main/java/classes/commands/Clear.java
#	src/main/java/classes/xml_manager/XMLMovieManager.java
#	src/main/java/helper/InputHandler.java
#	src/main/java/storage.xml

+++ src/main/java/exceptions/NotGreatThanException.java
@@ -1,9 +1,9 @@
 package exceptions;
 
 public class NotGreatThanException extends WarningException {
-    private final int bound;
+    private final long bound;
 
-    public NotGreatThanException(int bound) {
+    public NotGreatThanException(long bound) {
         this.bound = bound;
     }
